Correcting checks on valid AppliedZoneSpecs.
